fix: add edge case check for `testChainlogIntegrity`
Without this check, there would be a case where changes to the chainlog
not followed by a version bump would actually make the test pass.

Imagine this:

```
KEY A -> ADDRESS A
KEY B -> ADDRESS B
KEY C -> ADDRESS C
```

If the spell removes `KEY C` and adds `ANOTHER KEY` with `ADDRESS C` as a value, we'd get:

```
KEY A -> ADDRESS A
KEY B -> ADDRESS B
ANOTHER KEY -> ADDRESS C
```

This should cause a version update, but the test is not enforcing it.